+++ sdk/appconfiguration/azure-appconfiguration/tests/test_azure_appconfiguration_client.py
@@ -1132,7 +1132,7 @@ class TestAppConfigurationClient(AppConfigTestCase):
             self.client = client
             # prepare 200 configuration settings
             for i in range(200):
-                self.client.add_configuration_setting(
+                self.client.set_configuration_setting(
                     ConfigurationSetting(
                         key=f"sample_key_{str(i)}",
                         label=f"sample_label_{str(i)}",
@@ -1149,35 +1149,13 @@ class TestAppConfigurationClient(AppConfigTestCase):
                 page_etags.append(etag)
 
             # monitor page updates without changes
-            continuation_token = None
-            index = 0
-            request = HttpRequest(
-                method="GET",
-                url="/kv?key=sample_key_%2A&label=sample_label_%2A&api-version=2023-10-01",
-                headers={
-                    "If-None-Match": page_etags[index],
-                    "Accept": "application/vnd.microsoft.appconfig.kvset+json, application/problem+json",
-                },
+            has_changed = self.client.check_configuration_settings(
+                page_etags, key_filter="sample_key_*", label_filter="sample_label_*"
             )
-            first_page_response = self.client.send_request(request)
-            assert first_page_response.status_code == 304
-
-            link = first_page_response.headers.get("Link", None)
-            continuation_token = link[1 : link.index(">")] if link else None
-            index += 1
-            while continuation_token:
-                request = HttpRequest(
-                    method="GET", url=f"{continuation_token}", headers={"If-None-Match": page_etags[index]}
-                )
-                index += 1
-                response = self.client.send_request(request)
-                assert response.status_code == 304
-
-                link = response.headers.get("Link", None)
-                continuation_token = link[1 : link.index(">")] if link else None
+            assert has_changed == False
 
             # do some changes
-            self.client.add_configuration_setting(
+            self.client.set_configuration_setting(
                 ConfigurationSetting(
                     key="sample_key_201",
                     label="sample_label_202",
@@ -1198,39 +1176,11 @@ class TestAppConfigurationClient(AppConfigTestCase):
             assert page_etags[2] != new_page_etags[2]
 
             # monitor page after updates
-            continuation_token = None
-            index = 0
-            request = HttpRequest(
-                method="GET",
-                url="/kv?key=sample_key_%2A&label=sample_label_%2A&api-version=2023-10-01",
-                headers={
-                    "If-None-Match": page_etags[index],
-                    "Accept": "application/vnd.microsoft.appconfig.kvset+json, application/problem+json",
-                },
+            has_changed = self.client.check_configuration_settings(
+                page_etags, key_filter="sample_key_*", label_filter="sample_label_*"
             )
-            first_page_response = self.client.send_request(request)
-            # 304 means the page doesn't have changes.
-            assert first_page_response.status_code == 304
-
-            link = first_page_response.headers.get("Link", None)
-            continuation_token = link[1 : link.index(">")] if link else None
-            index += 1
-            while continuation_token:
-                request = HttpRequest(
-                    method="GET", url=f"{continuation_token}", headers={"If-None-Match": page_etags[index]}
-                )
-                index += 1
-                response = self.client.send_request(request)
-
-                # 200 means the page has changes.
-                assert response.status_code == 200
-                items = response.json()["items"]
-                for item in items:
-                    # Read the keys
-                    pass
-
-                link = response.headers.get("Link", None)
-                continuation_token = link[1 : link.index(">")] if link else None
+            # True means at least one page has changes.
+            assert has_changed == True
 
             # clean up
             self.tear_down()